requirements.txt
Flask==2.3.3
requests==2.31.0
python-whois==0.9.3
tldextract==5.1.1
gunicorn==21.2.0
